pyproject.toml
[project]
name = "rag-workflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "fastapi[standard]>=0.115.12",
    "fastembed>=0.7.1",
    "ipykernel>=6.29.5",
    "langgraph>=0.3.30",
    "openai>=1.74.0",
    "pydantic-settings>=2.8.1",
    "pymupdf>=1.26.4",
    "python-dotenv>=1.1.0",
    "qdrant-client>=1.13.3",
    "streamlit>=1.44.1",
    "unstructured>=0.17.2",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["app*"]